# examples/user-management/out/internal/api/handler_org_users.go
// Code generated by stego. DO NOT EDIT.

package api

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"reflect"
	"strconv"
	"strings"

	auth "github.com/example/user-management/out/internal/auth"
	slots "github.com/example/user-management/out/slots"
	"github.com/google/uuid"
)

// OrgUsersHandler handles HTTP requests for User entities.
type OrgUsersHandler struct {
	store                 Storage
	beforeCreateGate      slots.BeforeCreateSlot
	onEntityChangedFanOut slots.OnEntityChangedSlot
}

// NewOrgUsersHandler creates a new OrgUsersHandler.
func NewOrgUsersHandler(store Storage, beforeCreateGate slots.BeforeCreateSlot, onEntityChangedFanOut slots.OnEntityChangedSlot) *OrgUsersHandler {
	return &OrgUsersHandler{
		store:                 store,
		beforeCreateGate:      beforeCreateGate,
		onEntityChangedFanOut: onEntityChangedFanOut,
	}
}

// checkAncestors verifies that all ancestor entities in the URL hierarchy exist.
func (h *OrgUsersHandler) checkAncestors(w http.ResponseWriter, r *http.Request) bool {
	organizationID := r.PathValue("org_id")
	if organizationID == "" {
		handleError(w, r, BadRequest("missing org_id"))
		return false
	}
	organizationIDExists, organizationIDErr := h.store.Exists(r.Context(), "Organization", organizationID)
	if organizationIDErr != nil {
		handleError(w, r, InternalError("internal error"))
		return false
	}
	if !organizationIDExists {
		handleError(w, r, NotFound("Organization", organizationID))
		return false
	}
	return true
}

func (h *OrgUsersHandler) Create(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	{
		var rawMap map[string]json.RawMessage
		if err := json.Unmarshal(bodyBytes, &rawMap); err == nil {
			if kindRaw, ok := rawMap["kind"]; ok {
				var kind string
				if err := json.Unmarshal(kindRaw, &kind); err != nil || kind != "User" {
					handleError(w, r, BadRequest("kind must be User"))
					return
				}
			}
		}
	}
	var user User
	if err := json.Unmarshal(bodyBytes, &user); err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	user.ID = uuid.New().String()
	user.OrgID = r.PathValue("org_id")
	if h.beforeCreateGate != nil {
		metadataVal := ""
		if user.Metadata != nil {
			metadataVal = string(*user.Metadata)
		}
		slotReq := &slots.BeforeCreateRequest{
			Input: &slots.CreateRequest{
				Entity: "User",
				Fields: map[string]string{
					"email":        user.Email,
					"display_name": user.DisplayName,
					"role":         user.Role,
					"org_id":       user.OrgID,
					"metadata":     metadataVal,
				},
			},
			Caller: func() *slots.Identity {
				id := auth.IdentityFromContext(r.Context())
				return &slots.Identity{UserID: id.UserID, Role: id.Role, Attributes: id.Attributes}
			}(),
		}
		slotResult, slotErr := h.beforeCreateGate.Evaluate(r.Context(), slotReq)
		if slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
		if !slotResult.Ok {
			sc := http.StatusForbidden
			if slotResult.StatusCode > 0 {
				sc = int(slotResult.StatusCode)
			}
			handleError(w, r, errorForStatus(sc, slotResult.ErrorMessage))
			return
		}
		if slotResult.Halt {
			sc := http.StatusOK
			if slotResult.StatusCode > 0 {
				sc = int(slotResult.StatusCode)
			}
			w.WriteHeader(sc)
			return
		}
	}
	if err := h.store.Create(r.Context(), "User", user); err != nil {
		if errors.Is(err, ErrConflict) {
			handleError(w, r, Conflict("resource already exists"))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	if h.onEntityChangedFanOut != nil {
		if _, slotErr := h.onEntityChangedFanOut.Evaluate(r.Context(), &slots.OnEntityChangedRequest{Entity: "User", Action: "create"}); slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(presentEntity(user, "User", user.ID, "/api/user-mgmt/v1/organizations/"+r.PathValue("org_id")+"/users"+"/"+user.ID))
}

func (h *OrgUsersHandler) Read(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	id := r.PathValue("id")
	existing, err := h.store.Get(r.Context(), "User", id)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else {
			handleError(w, r, InternalError(err.Error()))
		}
		return
	}
	scopeData, err := json.Marshal(existing)
	if err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	var scopeMap map[string]any
	if err := json.Unmarshal(scopeData, &scopeMap); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	scopeVal, _ := scopeMap["org_id"].(string)
	if scopeVal != r.PathValue("org_id") {
		handleError(w, r, NotFound("User", id))
		return
	}
	var user User
	if err := json.Unmarshal(scopeData, &user); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(presentEntity(user, "User", id, "/api/user-mgmt/v1/organizations/"+r.PathValue("org_id")+"/users"+"/"+id))
}

func (h *OrgUsersHandler) Update(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	id := r.PathValue("id")
	existing, err := h.store.Get(r.Context(), "User", id)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	scopeData, err := json.Marshal(existing)
	if err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	var scopeMap map[string]any
	if err := json.Unmarshal(scopeData, &scopeMap); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	scopeVal, _ := scopeMap["org_id"].(string)
	if scopeVal != r.PathValue("org_id") {
		handleError(w, r, NotFound("User", id))
		return
	}
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	{
		var rawMap map[string]json.RawMessage
		if err := json.Unmarshal(bodyBytes, &rawMap); err == nil {
			if kindRaw, ok := rawMap["kind"]; ok {
				var kind string
				if err := json.Unmarshal(kindRaw, &kind); err != nil || kind != "User" {
					handleError(w, r, BadRequest("kind must be User"))
					return
				}
			}
		}
	}
	var user User
	if err := json.Unmarshal(bodyBytes, &user); err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	user.ID = id
	user.OrgID = r.PathValue("org_id")
	if err := h.store.Replace(r.Context(), "User", id, user); err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else if errors.Is(err, ErrConflict) {
			handleError(w, r, Conflict("resource already exists"))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	if h.onEntityChangedFanOut != nil {
		if _, slotErr := h.onEntityChangedFanOut.Evaluate(r.Context(), &slots.OnEntityChangedRequest{Entity: "User", Action: "update"}); slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(presentEntity(user, "User", id, "/api/user-mgmt/v1/organizations/"+r.PathValue("org_id")+"/users"+"/"+id))
}

func (h *OrgUsersHandler) Delete(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	id := r.PathValue("id")
	existing, err := h.store.Get(r.Context(), "User", id)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	scopeData, err := json.Marshal(existing)
	if err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	var scopeMap map[string]any
	if err := json.Unmarshal(scopeData, &scopeMap); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	scopeVal, _ := scopeMap["org_id"].(string)
	if scopeVal != r.PathValue("org_id") {
		handleError(w, r, NotFound("User", id))
		return
	}
	if err := h.store.Delete(r.Context(), "User", id); err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	if h.onEntityChangedFanOut != nil {
		if _, slotErr := h.onEntityChangedFanOut.Evaluate(r.Context(), &slots.OnEntityChangedRequest{Entity: "User", Action: "delete"}); slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
	}
	w.WriteHeader(http.StatusNoContent)
}

func (h *OrgUsersHandler) List(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	pageStr := r.URL.Query().Get("page")
	pageSizeStr := r.URL.Query().Get("pageSize")
	page, _ := strconv.Atoi(pageStr)
	if page < 1 {
		page = 1
	}
	pageSize, _ := strconv.Atoi(pageSizeStr)
	if pageSize < 1 {
		pageSize = 100
	}
	if pageSize > 65500 {
		pageSize = 65500
	}
	validFields := map[string]bool{"email": true, "display_name": true, "role": true, "org_id": true, "metadata": true}
	orderDir := r.URL.Query().Get("order")
	if orderDir != "" && orderDir != "asc" && orderDir != "desc" {
		handleError(w, r, BadRequest("invalid order value: "+orderDir+"; must be 'asc' or 'desc'"))
		return
	}
	var orderBy []OrderByField
	if orderByStr := r.URL.Query().Get("orderBy"); orderByStr != "" {
		for _, entry := range strings.Split(orderByStr, ",") {
			entry = strings.TrimSpace(entry)
			if entry == "" {
				continue
			}
			parts := strings.Fields(entry)
			fieldName := parts[0]
			dir := "asc"
			if len(parts) > 1 {
				d := strings.ToLower(parts[1])
				if d != "asc" && d != "desc" {
					handleError(w, r, BadRequest("invalid orderBy direction: "+parts[1]+"; must be 'asc' or 'desc'"))
					return
				}
				dir = d
			}
			if !validFields[fieldName] {
				handleError(w, r, BadRequest("invalid orderBy field: "+fieldName))
				return
			}
			orderBy = append(orderBy, OrderByField{Field: fieldName, Direction: dir})
		}
	} else {
		orderBy = []OrderByField{{Field: "created_time", Direction: "desc"}}
	}
	if orderDir != "" {
		for i := range orderBy {
			orderBy[i].Direction = orderDir
		}
	}
	var fields []string
	if fieldsStr := r.URL.Query().Get("fields"); fieldsStr != "" {
		hasID := false
		for _, f := range strings.Split(fieldsStr, ",") {
			f = strings.TrimSpace(f)
			if f == "" {
				continue
			}
			if f != "id" && !validFields[f] {
				handleError(w, r, BadRequest("invalid fields value: "+f))
				return
			}
			fields = append(fields, f)
			if f == "id" {
				hasID = true
			}
		}
		if !hasID {
			fields = append(fields, "id")
		}
	}
	searchExpr := r.URL.Query().Get("search")
	opts := ListOptions{Page: page, Size: pageSize, OrderBy: orderBy, Fields: fields, Search: searchExpr}
	scopeValue := r.PathValue("org_id")
	listResult, err := h.store.List(r.Context(), "User", "org_id", scopeValue, opts)
	if err != nil {
		if errors.Is(err, ErrSearch) {
			handleError(w, r, BadRequest(err.Error()))
		} else {
			handleError(w, r, InternalError(err.Error()))
		}
		return
	}
	w.Header().Set("Content-Type", "application/json")
	hrefBase := "/api/user-mgmt/v1/organizations/" + r.PathValue("org_id") + "/users"
	itemsSlice := reflect.ValueOf(listResult.Items)
	actualSize := itemsSlice.Len()
	presentedItems := make([]map[string]any, actualSize)
	for i := 0; i < actualSize; i++ {
		item := itemsSlice.Index(i).Interface()
		itemID := reflect.ValueOf(item).FieldByName("ID").String()
		itemData, err := json.Marshal(item)
		if err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		var apiItem User
		if err := json.Unmarshal(itemData, &apiItem); err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		presentedItems[i] = presentEntity(apiItem, "User", itemID, hrefBase+"/"+itemID)
		if len(fields) > 0 {
			allowed := map[string]bool{"id": true, "kind": true, "href": true}
			for _, f := range fields {
				allowed[f] = true
			}
			for k := range presentedItems[i] {
				if !allowed[k] {
					delete(presentedItems[i], k)
				}
			}
		}
	}
	result := map[string]any{
		"kind":  "UserList",
		"page":  page,
		"size":  actualSize,
		"total": listResult.Total,
		"items": presentedItems,
	}
	json.NewEncoder(w).Encode(result)
}

// OrgUsersPatchRequest contains the fields that can be partially updated.
type OrgUsersPatchRequest struct {
	DisplayName *string          `json:"display_name"`
	Role        *string          `json:"role"`
	Metadata    *json.RawMessage `json:"metadata"`
}

func (h *OrgUsersHandler) Patch(w http.ResponseWriter, r *http.Request) {
	if !h.checkAncestors(w, r) {
		return
	}
	id := r.PathValue("id")
	existing, err := h.store.Get(r.Context(), "User", id)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	existingData, err := json.Marshal(existing)
	if err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	var user User
	if err := json.Unmarshal(existingData, &user); err != nil {
		handleError(w, r, InternalError("internal error"))
		return
	}
	if user.OrgID != r.PathValue("org_id") {
		handleError(w, r, NotFound("User", id))
		return
	}
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	{
		var rawMap map[string]json.RawMessage
		if err := json.Unmarshal(bodyBytes, &rawMap); err == nil {
			if kindRaw, ok := rawMap["kind"]; ok {
				var kind string
				if err := json.Unmarshal(kindRaw, &kind); err != nil || kind != "User" {
					handleError(w, r, BadRequest("kind must be User"))
					return
				}
			}
		}
	}
	var patch OrgUsersPatchRequest
	if err := json.Unmarshal(bodyBytes, &patch); err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	if patch.DisplayName != nil {
		user.DisplayName = *patch.DisplayName
	}
	if patch.Role != nil {
		user.Role = *patch.Role
	}
	if patch.Metadata != nil {
		user.Metadata = patch.Metadata
	}
	if err := h.store.Replace(r.Context(), "User", id, user); err != nil {
		if errors.Is(err, ErrNotFound) {
			handleError(w, r, NotFound("User", id))
		} else if errors.Is(err, ErrConflict) {
			handleError(w, r, Conflict("resource already exists"))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	if h.onEntityChangedFanOut != nil {
		if _, slotErr := h.onEntityChangedFanOut.Evaluate(r.Context(), &slots.OnEntityChangedRequest{Entity: "User", Action: "patch"}); slotErr != nil {
			handleError(w, r, InternalError(slotErr.Error()))
			return
		}
	}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(presentEntity(user, "User", id, "/api/user-mgmt/v1/organizations/"+r.PathValue("org_id")+"/users"+"/"+id))
}